add some

+++ bin/install.sh
@@ -16,7 +16,11 @@ if [ "$?" != "0" ]; then
     fi
 fi
 curl -H "Cache-Control: no-cache" -O https://raw.githubusercontent.com/derrick921213/Derrick_package_manager-DPM-/main/dpm.py
-pyinstaller -F dpm.py
+pyinstaller -F dpm.py >/dev/null
+if [ "$?" != "0" ]; then
+    echo "Install Error!!"
+    exit 1
+fi
 echo [DPM] Install successful.
 sudo mv dist/dpm /usr/local/bin
 sudo rm -rf __pycache__ dist build dpm.spec dpm.py
